use meson build
This commit removes autotools in favour of meson build.  The way build time
configuration can be altered changed, and the INSTALL file is updated to
reflect that.

There are some other clean ups as well.  For example in charset directory
short names for replchars and specchars is no longer supported.  Microsoft
environment should not have any problem with long file name, so use them
everywhere.

The compat/ code is removed.  If someone really needs this stuff I promise
to fix issues as they arise, but quite frankly these compatibility things
appeared to refer to problems that do not exist on systems since year 2000.

If removal of config.h inclusion appears strange notice that project
arguments in main meson.build file adds -include directive that ensures
config.h is always included before anything else.  I think this is better.

One effect of this change is that catdoc can now be freely relocated with
-Dprefix=path argument and various charset referals in config.h, man page,
and wordview use the right path.  In same go the systemrc file is fixed to
do the same, and manual page path to that file is made literal as defined
when command is compiled.

As additional nice thing travis file is added to the project, that if
enabled will build the project automatically.  Assuming regression checks
are added travis can be made to run them, and that should make avoiding
obvious issues easy.

Signed-off-by: Sami Kerola <kerolasa@iki.fi>